+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   - Add contact page
   - Add version page
   - Add dashboard
+- **Admin Changes**
+  - Add admin dashboard
 - **Base Web Framework**
   - Rails for the web framework
   - Amazon AWS for hosting and continuous deployment